fixed float in quantity

+++ database/migrations/2017_11_17_172011_create_material_product_table.php
@@ -16,7 +16,7 @@ class CreateMaterialProductTable extends Migration
       Schema::create('material_product', function (Blueprint $table) {
             $table->integer('material_id')->unsigned();
             $table->integer('product_id')->unsigned();
-            $table->integer('quantity');
+            $table->float('quantity', 8, 4);
             $table->foreign('material_id')->references('id')->on('materials')->onDelete('cascade');
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
             $table->timestamps();
